Add login form component
